% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.deconv.R
\name{deconv}
\alias{deconv}
\title{Separate reads by genes and deconvolute them based on barcodes}
\usage{
deconv(fn, nRead = 1e+08, info.file, sample.IDs, Fprimer, Rprimer,
  primer.mismatch = 0, Find, Rind, index.mismatch = 0, gene,
  dir.out = NULL)
}
\arguments{
\item{fn}{Fully qualified name (i.e. the complete path) of the fastq file}

\item{nRead}{The number of bytes or characters to be read at one time. See 
\code{\link[ShortRead]{FastqStreamer}} for details}

\item{info.file}{Fully qualified name (i.e. the complete path) of the CSV 
file with the information needed on primers, indexes etc. (See details)}

\item{sample.IDs}{A character vector with the name of the column in info.file
containing the sample IDs}

\item{Fprimer, Rprimer}{A character vector with the name of the column in 
info.file containing the forward and reverse primer sequence, respectively}

\item{primer.mismatch}{The maximum number of primer mismatch}

\item{Find, Rind}{A character vector with the name of the column in info.file
containing the forward and reverse index sequence respectively}

\item{index.mismatch}{The maximum number of index mismatch}

\item{gene}{A character vector with the name of the column in info.file 
containing the name of the gene or other group idenifiers (see details)}

\item{dir.out}{The directory where to save the results. If NULL (default) 
then it will be set the same location where the input data was located}
}
\value{
A fastq file with the reads that were retained after removing the 
  indexes (with the suffix "_IndRm") and after removing the primers (with the
  suffix "_Ind_primerRm", in a folder named "Final") were end adaptor was 
  found (and removed) saved. A list with the total number of reads that were 
  processed and retained is also returned.
  
  When relevant, a text file with the sequence IDs that had multiple hits in 
  the preliminary search for the forward primer.
}
\description{
\code{deconv} takes a fastq file and will search for the forward primer and 
use this to separate the reads. That is, amplicons from multiple genes will 
be separated. Within each gene, reads are then separated based on forward 
and/or reverse index.  The end products are several fastq files - one for 
each samples, in as many folders as genes - where primers and indexes were 
removed.
}
\details{
This function applies only to reads with in-line indexes. That is, where the 
architecture of the reads is as follows:

F_index---F_primer---Target_sequence---R_primer---R_index

Note that the P7 adapter can be removed with \code{\link{rmEndAdaptor}}.

It is possible to control the number of mismatch, and IUPAC ambiguities codes
can be used in primers (i.e.the search for the primers is conducted with 
\code{fixed=FALSE}, which means (from Biostring): "an IUPAC ambiguity code in
the pattern can match any letter in the subject that is associated with the 
code, and vice versa". Note that indexes are searched with \code{fixed=TRUE}.

Information about the reads are passed with a comma separated file (CSV), 
whose path and name is passed with \code{info.file}. This must contain a 
column for each: the foward index, the reverse index, the forward primer, the
reverse primer, the sample IDs, and the gene (or an identifier of gene) being
amplified. The column headings are passed with the function arguments. While 
it is possible to include other columns where the users can record additional
information, these are effectively ignored. It is not mandatory to have both,
the forward and reverse indexes, but if one is not used, there is still the 
need to include a blank column in  \code{info.file} and indicate the column 
heading.

All sequences for indexes and primers are passed (as character vector) in 5' 
to 3' direction and are internally reversed and complemented when necessary.

\code{deconv} initially searches for the forward primer and separates the 
reads creating (if not existing already) a folder named as for the relevant 
information provided in the column \code{gene}, which is typically an 
identifier of the targeted genes. It is possible to use the column 
\code{gene} to group genes in other logical way than genes, but all identical
forward primer should have the same \code{gene} information. This is because 
(for efficiency) \code{deconv} uses only the first line for each forward 
primer to identify where the processed data should be saved and if multiple 
codes are used for \code{gene} for the same forward primer, these are 
actually ignored.

When relevant, a warning is reported and a text file with the sequence IDs 
that had multiple hits in the preliminary search for the forward primer is
saved.

After reads are separated based on the forward primer, indexes and primers 
are removed and processed samples are written to fastq files. Only reads 
where both primers (forward and reverse) and indexes (if there is information
for both in \code{info.file}) were found are retained.
}

