% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Detect a reference sequence}
\usage{
detect(data = NULL, dir.in = NULL, dir.out = NULL, ext = "fasta",
  ref_seq)
}
\arguments{
\item{data}{The output from \code{data.proc}}

\item{dir.in}{The directory where the fastq files are located. If NULL 
(default) an interactive window is used to select a folder}

\item{dir.out}{The directory where to save the results. If NULL (default) 
then \code{dir.out <- dir.in}}

\item{ext}{A character vector (of length=1) with the extension of the fasta
files. (Default "fasta")}

\item{ref_seq}{A character vector with the reference sequence}
}
\value{
A list where each element is an alignment of the sequences within 
  each sample with the reference sequence. The alignments and the number of 
  sequences found and mismatch is also written to file (see Details)
}
\description{
\code{detect} takes in either the output of \code{data.proc}, or a list of 
sequences from fasta files, and compare them with a reference sequence 
reporting the number of mismatch using \code{srdistance} from the package 
\code{shortRead}.
}
\details{
The output from \code{data.proc} can be passe with \code{data}. If no data is
passed to \code{detect}, then it will read fasta files in the directory 
passed with \code{dir.in}. If \code{dir.in=NULL}, then an interactive window 
will open to select the location of the files. As for \code{data.proc}, 
\code{detect} assumes that each file represents a sample.

A summary of the number of sequences found and the number of mismatch is 
returned as \code{data.frame} as well as being written to disk (Summary.csv) 
together with the alignments of the sequences provided with the reference 
sequence (in the folder "Final_alns"). The alignment is built using 
\code{PairwiseAlignments} from the package \code{Biostrings}.
}

