% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\name{fsc.bootstraps}
\alias{fsc.bootstraps}
\title{Compute bootstrap confidence intervals for estimated parameters with fsc}
\usage{
fsc.bootstraps(
  dir.in,
  nLoci = 10000,
  nSim = 100,
  maf = TRUE,
  ncpu = 0,
  nBatches = NULL,
  fsc.cmd = "fsc2702",
  fsc.path = "path",
  par.indLoci = "200000 0",
  par.nBlocks = 1,
  par.data = "DNA 100 0 2.5e-8 0.33",
  n = 1e+05,
  L = 50,
  nBoot = 1000,
  conf = 0.95,
  boot.type = "perc"
)
}
\arguments{
\item{dir.in}{The directory where the analysis was conducted}

\item{nLoci}{The number of polymorphic loci to retain}

\item{nSim}{The number of datasets that need to be simulated}

\item{maf}{Whether a MAF SFS (default) or a derived SFS is provided (if
\code{FALSE})}

\item{ncpu}{The number of CPU (threads) to use in the analysis. Automatically
handle if \code{ncpu=0} (default)}

\item{nBatches}{The number of batches (-B option)}

\item{fsc.path}{The path where fsc is installed or \code{"path"} if it is in
the PATH (that means that it can be called regardless of the working
directory)}

\item{par.indLoci}{The two integers value that need to be used for the number
of independent loci in the .par that it is used to run the simulations.}

\item{par.nBlocks}{The number of linkage blocks in the .par that it is used 
to run the simulations.}

\item{par.data}{The string to be used in the 'per Block: data type, num loci, 
rec. rate and mut rate + optional parameters' line of the .par that it is used 
to run the simulations. The default value is 
\code{par.data="DNA 100 0 2.5e-8 0.33"}}

\item{n}{The number of coalescent simulations to approximate the expected SFS
(-n option). This should be larger than 100,000.}

\item{L}{The number of optimization cycles (-L option). It should be >50}

\item{nBoot}{The number of bootstrap replicates}

\item{conf}{The confidence level to compute the confidence intervals}

\item{boot.type}{The method to be used to compute the confidence intervals. 
See \code{?boot.ci} for details. By default, percentile intervals are computed.}
}
\description{
Once fsc estimated parameter values, this function uses the *_maxL.par to
simulate data with the model parameters. From there, it estimates the
parameter values and confidence intervals are finally returned using the R
package \code{boot},
}
\details{
It is important that enough sites are simulated to ensure that sufficient
polymorphic loci are present in the simulated data. It is better to simulate
an excess of sites and retained those needed using \code{nLoci}.

Initial values are passed using the .pv so that a reduced numner of replicates can be passed
}
