% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{getnFiltered}
\alias{getnFiltered}
\title{Return the number of filtered sequences from a derep obj}
\usage{
getnFiltered(derep)
}
\arguments{
\item{derep}{}
}
\value{
A vector
}
\description{
Return the number of filtered sequences from a derep obj
}
\details{
This function is used internally by amplicR
}
