% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{getnUniques}
\alias{getnUniques}
\title{Return the number of unique sequences in a derep obj}
\usage{
getnUniques(derep)
}
\arguments{
\item{derep}{}
}
\value{
A vector
}
\description{
Return the number of unique sequences in a derep obj
}
\details{
This function is used internally by amplicR
}
