% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{getRow}
\alias{getRow}
\title{Extracts rows from a matrix into a vector}
\usage{
getRow(m, rn)
}
\arguments{
\item{m}{A matrix}

\item{rn}{The row number to extract}
}
\value{
A vector of same type as the matrix
}
\description{
The row rn becomes an element of a vector. This function is used internally by amplicR
}
