% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{getSeqFromDerep}
\alias{getSeqFromDerep}
\title{Extract sequences from a derep object}
\usage{
getSeqFromDerep(derepObj)
}
\arguments{
\item{derepObj}{The derep-element (that is one element from the list that constitutes the derep list)}
}
\value{
A character vector with the sequences
}
\description{
This function is used internally to batch-extract sequences from a derep object
}
