% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.phyloseq.R
\name{make.phyloseq}
\alias{make.phyloseq}
\title{Create a phyloseq object}
\usage{
make.phyloseq(dproc, sample.table = NULL, tax.table = NULL, phy.tree = NULL)
}
\arguments{
\item{dproc}{The \code{\link{data.proc}} output}

\item{sample.table}{The data.frame with sample metadata}

\item{tax.table}{A matrix with taxonomic assignments (see
\code{\link[phyloseq]{phyloseq}} for details)}

\item{phy.tree}{A phylogenetic tree (see
\code{\link[phyloseq]{phyloseq}} for details)}
}
\description{
\code{make.phyloseq} creates a \code{\link[phyloseq]{phyloseq}} object starting
  from \code{data.proc} output. Sample metadata and taxonomic assignments can
  be optionally included. If a minimum of two elements are passed to 
  \code{make.phyloseq}, then a phyloseq objext is returned, otherwise, 
  \code{data.proc} output will be returned as 
  \code{\link[phyloseq]{otu_table-class}}.
   
  
A \code{data.frame} can be passed with \code{sample.table} to include sample 
  metadata. In this case, the row names have to be identical to the sample 
  names included in the  \code{data.proc} output. The safest way to make sure
  this is the case is to use:
  \code{row.names(sampleTable) <- row.names(dproc$stable)} where 
  \code{sampleTable} is the  \code{data.frame} with the metadata. Note that 
  this assumes that the samples are in the same order as in the the sequence
  table.
}
\seealso{
\code{\link{data.proc}}, \code{\link[phyloseq]{phyloseq}}
}
