% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{ndada}
\alias{ndada}
\title{Return the number of unique sequences in a dada obj}
\usage{
ndada(dada_el)
}
\arguments{
\item{dada_el}{the element of a dada object}
}
\value{
A vector
}
\description{
Return the number of unique sequences in a dada obj
}
\details{
This function is used internally by amplicR
}
