% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{putna}
\alias{putna}
\title{Replace null values with NA}
\usage{
putna(x)
}
\arguments{
\item{x}{the element to evaluate}
}
\value{
An obj of same class and length as the input
}
\description{
Replace null values with NA
}
\details{
This function is used internally by amplicR
}
