% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{putzero}
\alias{putzero}
\title{Replace null values with zero or x with x - 1}
\usage{
putzero(x)
}
\arguments{
\item{x}{the element to evaluate}
}
\value{
An obj of same class and length as the input
}
\description{
Replace null values with zero or x with x - 1
}
\details{
This function is used internally by amplicR
}
