% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.deconv.R
\name{raw2data.proc}
\alias{raw2data.proc}
\title{From raw data to data.proc()}
\usage{
raw2data.proc(fn, nRead = 1e+08, rmEnd = FALSE, EndAdapter = "P7_last10",
  adapter.mismatch = 0, info.file, sample.IDs = "Sample_IDs",
  Fprimer = "F_Primer", Rprimer = "R_Primer", primer.mismatch = 0,
  Find = "F_ind", Rind = "R_ind", index.mismatch = 0, gene = "Gene",
  amplic.size = "Amplicon", truncQ = 2, qrep = FALSE, dada = TRUE,
  pool = FALSE, plot.err = FALSE, chim = TRUE, orderBy = "abundance")
}
\arguments{
\item{fn}{Fully qualified name (i.e. the complete path) of the fastq file}

\item{nRead}{The number of bytes or characters to be read at one time. See 
\code{\link[ShortRead]{FastqStreamer}} for details}

\item{rmEnd}{Whether \code{rmEndAdapter} should be performed (default: FALSE)}

\item{EndAdapter}{A character vector with the sequence of the end adapter,
"P7" or "P7_last10" (See details)}

\item{adapter.mismatch}{The maximum number of allowed mismatch (See details)}

\item{info.file}{Fully qualified name (i.e. the complete path) of the CSV 
file with the information needed on primers, indexes etc. (See details)}

\item{sample.IDs}{A character vector with the name of the column in info.file
containing the sample IDs}

\item{Fprimer}{A character vector with the name of the column in 
info.file containing the forward and reverse primer sequence, respectively}

\item{Rprimer}{A character vector with the name of the column in 
info.file containing the forward and reverse primer sequence, respectively}

\item{primer.mismatch}{The maximum number of primer mismatch}

\item{Find}{A character vector with the name of the column in info.file 
containing the forward and reverse index sequence respectively}

\item{Rind}{A character vector with the name of the column in info.file 
containing the forward and reverse index sequence respectively}

\item{index.mismatch}{The maximum number of index mismatch}

\item{gene}{A character vector with the name of the column in info.file 
containing the name of the gene or other group idenifiers (see details)}

\item{amplic.size}{A character vector with the name of the column in 
info.file containing the amplicon size of the PCR product}

\item{truncQ}{Truncate reads at the first instance of a quality score less 
than or equal to truncQ when conduction quality filtering. See
\code{\link[dada2]{fastqFilter}} for details}

\item{qrep}{Logical. Should the quality report be generated? (default 
\code{FALSE})}

\item{dada}{Logical. Should the dada analysis be conducted? (default 
\code{TRUE})}

\item{pool}{Logical. Should samples be pooled together prior to sample 
inference? (default \code{FALSE}). See \code{\link[dada2]{dada}} for
details}

\item{plot.err}{Logical. Whether error rates obtained from \code{dada} should
be plotted}

\item{chim}{Logical. Should the bimera search and removal be performed? 
(default \code{TRUE})}

\item{orderBy}{Character vector specifying how the returned sequence table 
should be sorted. Default "abundance". See 
\code{\link[dada2]{makeSequenceTable}} for details}
}
\value{
A list that has for elements the output of \code{\link{data.proc}} 
  for each PCR product
  
  Also, in addition to the output files described in the documentations for 
  \code{\link{rmEndAdapter}}, \code{\link{deconv}} and 
  \code{\link{data.proc}}, a text file, named "summary_nReads.txt" is saved 
  in the same location where the raw data are, summarising the number of 
  reads retained in each step of the analysis
}
\description{
This function is a wrapper for \code{\link{rmEndAdapter}}, 
\code{\link{deconv}} and \code{\link{data.proc}}. It takes in a raw fastq 
file, removes the end adapter, separates the reads based on their forward 
primers. Within each of the identified group, separates the reads based on 
barcodes (indexes) and eventually calls \code{\link{data.proc}} to process 
(quality checking, denoising and chimeras filtering) the retained data from 
the NGS run.
}
\details{
Note that the amplicon size for \code{\link{data.proc}} is obtained from the 
comma delimited file \code{info.file}, searching in the column with the 
heading indicated in \code{amplic.size}. Zeros can be used in this column if 
no truncation is wanted. For each entry in the column indicated with the 
argument \code{gene}, the function will use the first entry found in 
\code{amplic.size} for the relevant \code{gene}. If the same gene identifier 
is used for multiple forward primers, refer to the documentation for the 
\code{\link{deconv}} to see how multiple PCR product can be grouped together 
using the \code{gene} column).  Note that withing each gene, the same
amplicon length is used \code{raw2data.proc}. To use different amplicon sizes
within a gene, run the three functions (\code{\link{rmEndAdapter}},
\code{\link{deconv}} and \code{\link{data.proc}}) manually,
rather than with \code{raw2data.proc}.

By default, \code{dir.out} is set to the location where the input file is and
\code{verbose=FALSE}.

Please, see documentations for each functions for more information.
}

