% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{replaceSNP}
\alias{replaceSNP}
\title{Replace SNPs with IUPAC ambiguities}
\usage{
replaceSNP(genotypes, seqAlleles, SNPpositions)
}
\arguments{
\item{genotypes}{(named integer) The genotypes (0,1,2) named with SNPs (see
Details)}

\item{seqAlleles}{possible sequences of the alleles or baseAllele (see \code{make.alleles} for details)}

\item{SNPpositions}{(integer) The position where the SNPs occur}
}
\description{
Replace SNPs with IUPAC ambiguities
}
\details{
See \code{make.alleles} for details on how \code{genotypes} should be formatted
}
