% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{sel.match}
\alias{sel.match}
\title{Select appropriate hits when searching a patter}
\usage{
sel.match(i, subjectSet, patt_hits, patt, type = "F")
}
\arguments{
\item{i}{The index to subset patt_hitts}

\item{subjectSet}{The set of sequences that were used for the search}

\item{patt_hits}{The output from Biostrings::vmatchPattern()}

\item{patt}{The patter used for the search}

\item{type}{The type of match, either forward "F", or reverse "R"}
}
\description{
This function is used internally by amplicR
}
