% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{subsetDerep}
\alias{subsetDerep}
\title{Subset a derep-class based on abundance of the reads}
\usage{
subsetDerep(derep, minAbund = 2)
}
\arguments{
\item{derep}{The derep object to subset}

\item{minAbund}{The threshold below which reads are removed}
}
\value{
A list where the first two elements are as a derep-class, and the third is rmed. See details.
}
\description{
This function retains the reads that have their abundance >= \code{minAbund}.
It returns an object similar to a derep-class in its structure. Indeed it has
a \code{$uniques} and \code{quals} elemnt for the retained sequences, but,
rather than having \code{map} as third element in the list, it has
\code{rmed}, which is a vector with the index of the positions of the
\code{uniques} from the original object that were removed.
}
