% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2fasta.R
\name{table2fasta}
\alias{table2fasta}
\title{Write unique sequences to a fasta file for each sample where sequence names  
match the sequence table and include abundance data}
\usage{
table2fasta(stable, seq.list = NULL, dir.out = NULL)
}
\arguments{
\item{stable}{The sequence table (output of \code{dada2::makeSequenceTable})
or \code{data.proc}}

\item{seq.list}{The seq_list element from \code{data.proc}. Required if the 
input table is the output from \code{data.proc}. If NULL (default), it is 
assumed that the input table is from \code{dada2::makeSequenceTable} and 
sequences are named with a seq# pattern, where # is a sequential number 
respect to the sequence in the table.}

\item{dir.out}{The directory where to save fasta files}
}
\value{
A fasta file for each sample
}
\description{
\code{table2fasta} takes the output from 
\code{\link[dada2]{makeSequenceTable}} or the table element from 
\code{data.proc} as input and generates a fasta file for each sample with 
unique sequences. It uses internally \code{\link[dada2]{uniquesToFasta}}. 
However, sequences are named with a sequential number, which is consistent 
across all samples and abundance data is appended to the sequence names.
}
\details{
The input can be either the \code{\link[dada2]{makeSequenceTable}} or 
\code{data.proc} output. Note that \code{data.proc} output is actually a 
modified \code{\link[dada2]{makeSequenceTable}} output where the column 
names, which normally are the actual sequeces, have been replaced for 
convinience. \code{data.proc} reports the sequence IDs and matching sequences 
in a separated elelment of the returned list. If the input is 
from \code{data.proc}, then also the \code{seq_list} element needs to be 
passed to \code{table2fasta}.

In each file, sequences are named with a sequential number, which is 
consistent across all samples. For example, if sample1 has sequence seq1 and 
seq2 and sample2 has seq1 and seq3, the sequences in the fasta file for 
sample1 will be named seq1 and seq2, and those in the fasta file for sample2 
will be named seq1 and seq3, so that it will be possible to match back the 
same sequences in different samples in down stream analysis.
}
\seealso{
\code{\link{data.proc}},
  \code{\link[dada2]{uniquesToFasta}}, \code{\link[dada2]{makeSequenceTable}}
}

