% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{where2trim}
\alias{where2trim}
\title{Select position where a pattern has been found}
\usage{
where2trim(mismatch, subjectSet, patt_hits, patt, type = "F")
}
\arguments{
\item{mismatch}{The number of allowed mismatches}

\item{subjectSet}{The set of sequences that were used for the search}

\item{patt_hits}{The output from Biostrings::vmatchPattern()}

\item{patt}{The patter used for the search}

\item{type}{The type of match, either forward "F", or reverse "R"}
}
\value{
A list with start positions
}
\description{
Takes a set of sequences and the output from
\code{Biostrings::vmatchPattern()} and identify the first positions after the
match with a pattern
}
\details{
This function is used internally by amplicR
}
