% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate.seqs.R
\name{collate.seqs}
\alias{collate.seqs}
\title{Collate sequences}
\usage{
collate.seqs(ldproc = NULL, rdas.in = NULL, dir.out = NULL)
}
\arguments{
\item{ldproc}{If the data are already in R memory, a list whose element are
\code{data.proc} outputs.}

\item{rdas.in}{If \code{ldproc=NULL}, a character vector with the path to the 
rda files where the outputs from data.proc that have to be collated were 
saved.}

\item{dir.out}{The directory where to save the results. If NULL (default) 
then it will be set to the working directory}
}
\value{
Return a list with several elements:
  
  \itemize{ 
  \item $lbysamples: A list whose elements are vectors with the unique 
        sequences (names) for each sample and their abundance (values). 
  \item $stable: The sequence table 
  \item $seq_list: The sequences and matching
  sequences IDs 
  \item $call: The function call}
  
  Several files are also saved to disk, these include:
  \itemize{ 
  \item Seq_table.csv Sequence table
  \item Seq_list.csv List of sequences and their matching IDs
  \item collated.rda R data file containing the list returned by collate.seqs 
     (see above)
     }
}
\description{
\code{collate.seqs} can be used to collate together outputs from 
\code{data.proc}.
}
\details{
This function was developed because, if, for example, sequencing runs are 
carried out for the same samples, different  \code{data.proc} outputs will 
also exists and you may want to collate together all the sequences from the 
same sample. \code{collate.seqs} does exactly this. By passing either the 
path to, or a list of, \code{data.proc} outputs, \code{collate.seqs} collate 
together the sequences.

\code{collate.seqs} will generate an output that it is very similar to
\code{data.proc} but with the combined seqs for the same samples.
}
\examples{
# Select the directory where the example data are stored
example.data <- system.file("extdata", "HTJ", package="amplicR")
# Select a temporary directory where to store the outputs
out <- tempdir()

# Run data.proc with bp=140 and save in a sub folder
out140 <- paste(out, "out140", sep="/")
HTJ.140 <- data.proc(example.data, out140, bp=140)

# Make up a some new sequence to collate by running data.proc with bp=141 and 
# saving in a sub folder
out141 <- paste(out, "out141", sep="/")
HTJ.141 <- data.proc(example.data, out141, bp=141)

# Collate seqs of the two objects HTJ.140 and HTJ.141 
collate.seqs(ldproc=list(HTJ.140, HTJ.141), dir.out=out)

# As above but using the .rda files
rdas <- paste(out, c("out140", "out141"), sep="/")
collate.seqs(rdas.in=rdas, dir.out=out)

# Clean up the temp directory
unlink(out, recursive=TRUE)
}

