% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Detect a reference sequence}
\usage{
detect(data = NULL, rda.in = NULL, dir.out = NULL, ref_seqs)
}
\arguments{
\item{data}{The output from \code{data.proc}}

\item{rda.in}{The fully qualified (i.e. including the path) name of the .rda 
file where the output from \code{data.proc} is saved (it assumes the object
name was not changed from the default \code{dproc})}

\item{dir.out}{The path where to save the results. If NULL and data is also 
NULL, the directory where the .rda file is located is used. If no file path
is provided, an interactive windows is used to select the folder}

\item{ref_seqs}{A \strong{named} character vector with the reference 
sequence(s)}
}
\value{
A \code{list} with two elements:
  \itemize{
    \item $detect_results A list with a result table, for each reference 
           sequence, with the minimum mismatch count
    \item $alns A list with an alignment, for each reference sequence as 
           elements, of the sequences in the sequence table with the 
           reference sequence
    \item $detect_seq_list A list where elements (one for each reference 
           sequence) are a \code{data.frame} with the sequence IDs and the 
           minimum number of differences with the reference sequence       
    }

  These results are also written to text files
}
\description{
\code{detect} takes in either the output of \code{data.proc}, or load it up 
from a .rda file, and compare the sequences with a reference sequence 
reporting the number of mismatch using \code{srdistance} from the package 
\code{shortRead}.
}
\details{
The output from \code{data.proc} can be passed with \code{data}. If no data 
is passed to \code{detect}, then it will load the .rda file passed with 
\code{rda.in}. If \code{rda.in=NULL}, then an interactive window will open to
select the location of the file.

If both \code{dir.out=NULL} and \code{rda.in=NULL}, then the path where to 
save the results will be asked with an interactive window. If the .rda file 
path is provided, then the folder where .rda is located will be selected as
output folder.

A summary of the number of sequences found and the minimum number of mismatch
within each sample is returned as \code{matrix} with the same layout as the 
sequence table and a \code{data.frame} with the sequence IDs and the 
number of mismatch. There will be as many tables as the length of the character 
vector passed with \code{ref_seqs}. These results are also written to disk as
text files along with the alignments of the sequences provided with the 
reference sequence (in the folder "Final_alns"). The alignments are built using
\code{PairwiseAlignments} from the package \code{Biostrings}.
}
\examples{
# Select the directory where the example data are stored
example.data <- system.file("extdata", "HTJ", package="amplicR")
# Select a temporary directory where to store the outputs
out <- tempdir()
# Process raw data
HTJ.test <- data.proc(example.data, out, bp=140)
# Referece Mycobacteriumavium subspecies paratuberculosis sequence 
HTJ <- "CTGCGCGCCGGCGATGACATCGCAGTCGAGCTGCGCATCCTGACCAGCCGACGTTCCGATCTGGTGGCTGATCGGACCCGGGCGATCGAACCGAATGCGCGCCCAGCTGCTGGAATACTTTCGGCGCTGGAACGCGCCTT"
# Naming the reference sequence
names(HTJ) <- "HTJ" 

# Use 'detect' to verify the presence of Mycobacteriumavium subspecies 
# paratuberculosis
det <- detect(HTJ.test, dir.out=out, ref_seqs=HTJ)

# Clean up the temp directory
unlink(out, recursive=TRUE)
}
\seealso{
\code{\link{data.proc}}, \code{\link[shortRead]{srdistance}}, 
  \code{\link[Biostrings]{PairwiseAlignments}}
}

